refactor(server): move share preview renderer into package
