+++ packages/doc-server/src/snapshot.ts
@@ -5,7 +5,7 @@ import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import { getDocumentBySlug } from '../../../server/db.js';
 import { getCanonicalReadableDocumentSync } from '../../../server/collab.js';
 import { recordSnapshotPublish } from '../../../server/metrics.js';
-import { buildSharePreviewModel, renderSharePreviewHtmlPage, resolvePublicOrigin } from '../../../server/share-preview.js';
+import { buildSharePreviewModel, renderSharePreviewHtmlPage, resolvePublicOrigin } from './share-preview.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
